refactor: move MCP_TOOL_CALL_TIMEOUT_MS to shared constants
Moves the timeout constant from the route layer (proxy.ts) to
core/constants.ts to fix circular dependency (proxy ↔ virtual-mcp)
and improper layering (mcp-clients importing from routes).
